[DEV] Fetching per commands (#56)

+++ src/commands/fun/poll.ts
@@ -1,13 +1,13 @@
 import { Command, CommandMessage, Description } from '@typeit/discord';
 import { Logger } from 'Services/logger.service';
 import { PollQuestion, selectionEmojis } from 'Types/poll';
-import { environment } from 'Utils/environment';
+import Utility from 'Utils/utility';
 import * as chalk from 'chalk';
 import { ClientUser, EmbedField, Message, MessageEmbed } from 'discord.js';
-
 export class Poll {
   private logger: Logger;
   private alphabet: string[] = [...'abcdefghijklmnopqrstuvwxyz'];
+
   constructor() {
     this.logger = new Logger();
   }
@@ -48,20 +48,23 @@ export class Poll {
   private async createPollingObject(
     command: CommandMessage
   ): Promise<Message | void> {
+    if (command.deletable) command.delete();
+
     const commandString = command.content.split('poll ').pop();
     const pollArray = commandString
-      ?.split(/[\[\]']+/g)
+      ?.split(/[[\]']+/g)
       .filter((e) => e.trim() != '');
 
     if (!pollArray?.length) {
       this.logger.error(
         `${chalk.bold('BOT ERROR')}: incorrect formatting used on command`
       );
-      return Promise.reject();
-    }
-
-    if (command.deletable) {
-      command.delete();
+      return Utility.sendMessage(
+        command,
+        '**Please use correct formatting**',
+        'channel',
+        5000
+      );
     }
 
     const poll = new PollQuestion(pollArray[0], pollArray.slice(1));
@@ -76,8 +79,19 @@ export class Poll {
   @Command('poll')
   @Description('Create a poll for friends to answer (max 26 options)')
   async init(command: CommandMessage): Promise<Message | void> {
-    return this.createPollingObject(command).catch(() => {
-      return command.reply(environment.error);
-    });
+    try {
+      return this.createPollingObject(command);
+    } catch (e: unknown) {
+      if (command.deletable) await command.delete();
+      this.logger.error(`Command: 'poll' has error: ${(e as Error).message}.`);
+      return Utility.sendMessage(
+        command,
+        `The following error has occurred: ${
+          (e as Error).message
+        }. If this error keeps occurring, please contact support.`,
+        'channel',
+        5000
+      );
+    }
   }
 }
